% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_split_expressions.R
\name{get_split_expressions}
\alias{get_split_expressions}
\title{Get split conditions for edges in tree}
\usage{
get_split_expressions(h2o_tree_R)
}
\arguments{
\item{h2o_tree_R}{a sinlge \code{data.frame} output from
\code{H2OTreeConvertR::h2o_tree_convertR}.}
}
\value{
\code{data.frame} containing left and right split conditions for
parent nodes.
}
\description{
This function manipulates the output of
\code{H2OTreeConvertR::h2o_tree_convertR} to get the split condition for each
edge in a tree
}
\examples{
library(h2o)

h2o.init()

prostate.hex = h2o.uploadFile(path = system.file("extdata",
                                                 "prostate.csv",
                                                 package = "h2o"),
                              destination_frame = "prostate.hex")

prostate.hex$RACE <- as.factor(prostate.hex$RACE)

prostate.hex$DPROS <- as.factor(prostate.hex$DPROS)

prostate.gbm = h2o.gbm(x = c("AGE", "RACE", "DPROS", "DCAPS", "PSA", "VOL", "GLEASON"),
                       y = "CAPSULE",
                       training_frame = prostate.hex,
                       ntrees = 5,
                       max_depth = 5,
                       learn_rate = 0.1)

h2o_trees <- H2OTreeConvertR::h2o_tree_convertR(prostate.gbm)

tree_1_split_expr <- get_split_expressions(h2o_trees[[1]])

}
