% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gv_to_table.R
\name{gv_to_table}
\alias{gv_to_table}
\title{Convert h2o tree .gv file to tabular structure}
\usage{
gv_to_table(gv_file, detail)
}
\arguments{
\item{gv_file}{.gv file file including path to import.}

\item{detail}{has the gv file been processed with the option --detail?}
}
\value{
\code{data.frame} containing tree structure, if the gv file is of an
empty tree then \code{NULL} is returned.
}
\description{
Reads a single .gv file containing a tree from a h2o gbm or drf and parses it
into a tabular structure
}
