% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_h2o_encoding.R
\name{map_h2o_encoding}
\alias{map_h2o_encoding}
\title{Get temrinal node paths in terms of L/R directions and variable conditions}
\usage{
map_h2o_encoding(h2o_trees_R)
}
\arguments{
\item{h2o_trees_R}{\code{list} output from the
\code{H2OTreeConvertR::h2o_tree_convertR} function.}
}
\value{
\code{list} containing split conditions to reach terminal nodes for
each tree.
}
\description{
This function takes the output from \code{H2OTreeConvertR::h2o_tree_convertR}
and for each tree returns the terminal node decision paths in terms of L/R
directions and variable conditions. This can be used to map the output from
\code{h2o::h2o.predict_leaf_node_assignment} to variable condtions.
}
\examples{
library(h2o)

h2o.init()

prostate.hex = h2o.uploadFile(path = system.file("extdata",
                                                 "prostate.csv",
                                                 package = "h2o"),
                              destination_frame = "prostate.hex")

prostate.hex$RACE <- as.factor(prostate.hex$RACE)

prostate.hex$DPROS <- as.factor(prostate.hex$DPROS)

prostate.gbm = h2o.gbm(x = c("AGE", "RACE", "DPROS", "DCAPS", "PSA", "VOL", "GLEASON"),
                       y = "CAPSULE",
                       training_frame = prostate.hex,
                       ntrees = 5,
                       max_depth = 5,
                       learn_rate = 0.1)

h2o_trees <- H2OTreeConvertR::h2o_tree_convertR(prostate.gbm)

h2o_terminal_node_encoding_mapping <- map_h2o_encoding(h2o_trees)

}
